Remove User and Security Implementation and improved Food Controller and Service

+++ src/main/java/com/petcc/restaurante/service/FoodService.java
@@ -1,11 +1,9 @@
 package com.petcc.restaurante.service;
 
 import java.util.List;
-import java.util.stream.Collectors;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import com.petcc.restaurante.model.Food;
 import com.petcc.restaurante.repository.FoodRepository;
 
@@ -14,21 +12,24 @@ public class FoodService {
     @Autowired
     private FoodRepository foodRepository;
     public List<Food> getAllFoods(){
-        return foodRepository.findAll().stream().collect(Collectors.toList());
+        return foodRepository.findAll();
     }
     public Food saveFood(Food food){
-        Food savedFood = foodRepository.save(food);
-        return savedFood;
+        return foodRepository.save(food);
+       
     }
-    public void deleteFood(Long id){
-        foodRepository.deleteById(id);
+    public boolean deleteFood(Long id){
+        if (foodRepository.existsById(id)) {
+            foodRepository.deleteById(id);
+            return true;
+        }
+        return false;
     }
     public Food updateFood(Long id, String title, double price, String image){
         Food food = foodRepository.findById(id).orElseThrow(() -> new RuntimeException("Food not found"));
         food.setTitle(title);
         food.setPrice(price);
         food.setImage(image);
-        foodRepository.save(food);
-        return food;
+        return foodRepository.save(food);
     }
 }
